% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create42Token.R
\name{create42TokenfromJSON}
\alias{create42TokenfromJSON}
\title{Create an Oauth2.0 token for 42 API using a credential files}
\usage{
create42TokenfromJSON(credentials)
}
\arguments{
\item{credentials}{String - path to your credentials JSON file}
}
\value{
A Token2.0 reference class (RC) object, specific to 42.
}
\description{
See \link[R42Api]{create42Token} to know what variables are needed.
}
